+++ jest.config.js
@@ -4,9 +4,7 @@
 const config = require('@polkadot/dev/config/jest.cjs');
 
 module.exports = Object.assign({}, config, {
-  automock: false,
   browser: true,
-  clearMocks: true, // Automatically clear mock calls and instances between every test
   moduleNameMapper: {
     '@polkadot/extension-(base|chains|dapp|inject|ui)(.*)$': '<rootDir>/packages/extension-$1/src/$2',
     // eslint-disable-next-line sort-keys
